+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-## [1.0.6] - 2010-08-22
+## [1.1] - 2020-09-03
+### Added
+- New `GridView` class.
+- Confirmation message on actions.
+- Search input with support for relational fields.
+- Sort table view columns
+- More UI components for table and grid view.
+  - Link
+  - Icons
+
+## [1.0.6] - 2020-08-22
 - Fixed typos
 - Fixed filter stub file name